Split the main.tsx to ce/ee/deploy
