+++ packages/admin-server/src/views/main.pug
@@ -31,4 +31,4 @@ html
     script(src=`${staticPath}js/polyfill.js`)
     script(src=`${staticPath}js/firebase-app.js`)
     div(id='root')
-      script(src= `${staticPath}hub.js`)
+      script(src= `${staticPath}main.js`)
