+++ packages/client/src/main-ce.tsx
@@ -1,0 +1,186 @@
+import * as React from 'react';
+import ReactDOM from 'react-dom';
+import {ApolloProvider} from 'react-apollo';
+import {genClient} from 'canner/lib/components/index';
+import GraphqlClient from 'canner-graphql-interface/lib/graphqlClient/graphqlClient';
+import {LocalStorageConnector} from 'canner-graphql-interface';
+import {notification, Button} from 'antd';
+import {BrowserRouter, Route} from 'react-router-dom';
+import {BackgroundTokenSyncer} from './workers/backgroundTokenSyncer';
+import Main, { MainPageSidebarItem } from 'containers/mainPage';
+import { appPrefix } from 'utils/env';
+
+// Icons
+import iconJupyterHub from 'images/icon-jupyterhub.svg'
+
+// Components
+import Jupyterhub from 'containers/jupyterhubPage';
+
+const graphqlClient = new GraphqlClient({
+  uri: (window as any).graphqlEndpoint,
+  fetch: (uri, options) => {
+    const token = window.localStorage.getItem('canner.accessToken');
+    options.headers = {
+      Authorization: `Bearer ${token}`,
+      ...options.headers || {}
+    };
+    return fetch(uri, options);
+  },
+});
+
+const fakeData = {
+  me: {
+    groups: [{
+      id: 'groupId1',
+      name: 'group1',
+      displayName: 'c-Group 1',
+      instanceTypes: [{
+        id: 'g-it1',
+        name: 'IT1',
+        displayName: 'group1 it'
+      }],
+      images: [{
+        id: 'g-it1',
+        name: 'IT1',
+        displayName: 'group1 im',
+      }]
+    }, {
+      id: 'groupId2',
+      name: 'group2',
+      displayName: 'Group 2',
+      instanceTypes: [{
+        id: 'ggit1',
+        name: 'IT1',
+        displayName: 'group2 it'
+      }],
+      images: [{
+        id: 'ggit2',
+        name: 'IT1',
+        displayName: 'group2 im',
+      }]
+    }, {
+      id: 'everyone',
+      name: 'everyone',
+      displayName: 'Group DisplayName',
+      instanceTypes: [{
+        id: 'everyone-it',
+        name: 'it',
+        displayName: 'gpu0',
+        gpuLimit: 0,
+        cpuLimit: 0.5,
+        memoryLimit: 4,
+      }, {
+        id: 'everyone-it2',
+        name: 'it',
+        displayName: 'gpu1',
+        gpuLimit: 1
+      }],
+      images: [{
+        id: 'everyone-image',
+        name: 'b-cpu',
+        displayName: 'b-cpu',
+        type: 'cpu'
+      }, {
+        id: 'everyone-image2',
+        name: 'a-gpu',
+        displayName: 'a-gpu',
+        type: 'gpu'
+      }, {
+        id: 'everyone-image3',
+        name: 'c-img',
+        displayName: 'c-img',
+        type: 'both'
+      }]
+    }]
+  },
+}
+
+const connector = new LocalStorageConnector({
+  defaultData: fakeData,
+  localStorageKey: 'infuse-hub'
+})
+
+const client = genClient(process.env.NODE_ENV === 'production' ?
+  {graphqlClient} :
+  {
+    connector,
+    schema: {
+      me: {type: 'object'},
+    }
+  });
+
+class Hub extends React.Component {
+  render() {
+    const sidebarItems: MainPageSidebarItem[] = [
+      {
+        title: 'JupyterHub',
+        subPath: 'hub',
+        icon: iconJupyterHub,
+      },
+    ]
+
+    return (
+      <BrowserRouter>
+        <ApolloProvider client={client}>
+          <Main sidebarItems={sidebarItems}>
+            {/* Jupyterhub */}
+            <Route path={`${appPrefix}g/:groupName/hub`} exact>
+              <Jupyterhub />
+            </Route>
+          </Main>
+        </ApolloProvider>
+      </BrowserRouter>
+    )
+  }
+}
+
+/**
+ * Background worker
+ */
+function checkStatus(response) {
+  if (response.status >= 200 && response.status < 300) {
+    return response;
+  } else {
+    const error = new Error(response.statusText);
+    (error as any).response = response;
+    throw error;
+  }
+}
+
+function parseJSON(response) {
+  return response.json();
+}
+
+const tokenSyncWorker = new BackgroundTokenSyncer({
+  appPrefix: (window as any).APP_PREFIX,
+  refreshTokenExp: (window as any).refreshTokenExp,
+  accessTokenExp: (window as any).accessTokenExp,
+  getNewTokenSet: () => {
+    return fetch(`${(window as any).APP_PREFIX}oidc/refresh-token-set`, {
+      method: 'POST'
+    })
+    .then(checkStatus)
+    .then(parseJSON);
+  },
+  reLoginNotify: ({loginUrl}) => {
+    // notify with fixed card
+    notification.warning({
+      message: 'Warning',
+      description: 'In less than 1 minute, you\'re going to be redirected to login page.',
+      placement: 'bottomRight',
+      duration: null,
+      btn: (
+        <Button type="primary" onClick={() => window.location.replace(`${(window as any).APP_PREFIX}oidc/logout`)}>
+          Login Again
+        </Button>
+      ),
+      key: 'refreshWarning'
+    });
+  }
+})
+tokenSyncWorker.run().catch(console.error);
+
+// render
+ReactDOM.render(
+  <Hub />
+, document.getElementById('root'));
