+++ packages/client/tslint.json
@@ -1,3 +1,3 @@
 {
-  "extends": ["../../tslint.json", "tslint-react"]
+  "extends": ["../../tslint.json"]
 }
